requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
django-graphql-jwt==0.3.1
flake8==3.8.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
mccabe==0.6.1
promise==2.3
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.0
pytz==2021.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
